#3042 FIX Fixed configuring the sender of mails (Text and HTML mails)
It is now really possible to configure the sender name and address
of notifications using the rule based notifications. This was not
completely working as sendmail was called with the wrong arguments
(-f instead of -F).

The from address can now be configured like this:

Company Mointoring <monitoring@example.com>

Tested with sendmail compatibility version from postfix.