#3171 FIX Logout in one site will log out of all sites sharing the same auth.secret
Site that share the file <tt>etc/auth.secret</tt> also share the same login.
A user logging in in one of these sites is automatically logged in in all
others. But logging out did only work when the logout was done on the same
site as the login. Now you can log in at one site and log out at the other.
Note: Also <tt>omd cp</tt> creates sites with the same secret.