#2991 Snapshot restore: Improved robustness and logging
The restore operation tries to complete the snapshot extraction with the rest of
the data when it encounters critical errors in the crucial extract phase.
Error information is logged in the file <tt>~/var/log/web.log</tt>.
This should never happen anyway - there are several mechanisms to prevent this.