3086 FIX fixed incorrect host list when one wato folder was a substring of another
When a view is filtered by a wato folder, that search was a pure prefix match, so if a view was
opened for the folder "abc" it would also contain all hosts from the folder "abcde".
This affected (probably among other things) the "Virtual Host Tree" and the "Tree of Folders"
snapins.