3364 FIX Fixed exception when creating a view of a single log entry
When one creates a view of a single log entry, this resulted in an
exception like "TypeError: 'NoneType' object is not iterable". This
has been fixed now. However it is currently not possible to create
such a view since the logfile info (data source) is not supporting
referencing of single log entries.