3371 FIX Be compatible to old site=None setting read from hosts.mk
Old WATO versions saved the site attribute with a value of None which
is not allowed in current WATO versions. Either the key has to be missing
when no specific site is configured or the key has to be missing. Cleaned
this up.

Fixed exception: TypeError (cannot concatenate 'str' and 'NoneType' objects)