3329 FIX Function discover_services now allows the URL parameter mode
The <tt>mode</tt> can e.g. be set to <tt>fixall</tt> in order to remove execeeding
and add missing services. The implementation was always present, but the validation
of the request complained that <tt>mode</tt> was not a valid parameter. This has
been fixed. Thanks to Thomas Huber.