3117 FIX windows agent: fixed installer may fail to re-start service on update
When updating the agent (i.e. through the auto updater) there was only a fairly short timeout for
the agent-service to stop.
If the agent failed to stop in time (i.e. because of a long-running plugin) the agent was still
installed but the service wasn't started afterwards.
This has now been fixed by enforcing an unlimited wait when the service is being shut down.