Horrible temporary hack needed to make cmc work again.
The ugly #ifdefs should really die soon, but the alternative is currently
even worse: Duplicating headers and classes in cmc even more.  What *really*
should be done is putting a thin compatibility layer into the system,
abstracting away from the differences in the Nagios/CMC worlds.  This isn't
hard, just a bit of mechanical work, but this needs to be postponed for
now...  :-/