3424 FIX ipmi: fixed missing compact/discrete sensor data
To get this check work you have to install the
check_mk_agent.linux.

The agent now provides an additional section for
compact/discrete sensors. Older ipmi check versions
do not handle this new section, but are still compatible.
To use this feature you have to upgrade the ipmi check
to the new version.