3449 The host alias can now be used in BI aggregates
When matching for hosts in BI aggregations it is now possible to match
hosts with regular expressions by their host alias instead of the host name.

Additionally it is now possible to use the host alias in the titles of the
aggregations.

Instead of using $1$ for matchin the host name or the host alias, you can now
use the variables $HOSTNAME$ or $HOSTALIAS$ to use them in your BI rules.