Fixed the event daemon protocol, its implementation and its usage.
socket.recv can (and will, see the POSIX spec!) sometimes result in a
partial read, so we have to handle this, and basically the same holds for
socket.send.  As a consequence, our protocol needs a way to determine when a
request is complete.  One way of doing this is shutting down the writing
side, but this would make it impossible to handle multiple requests over the
same connection.

The solution on the receiving side is simple: Do it like Livestatus and
terminate a request with an empty line.  A shutdown will of course continue
to terminate a request, too.  On the sending side, use socket.sendall
instead of socket.send.

In a nutshell: 99.9% of all uses of recv/send/read/write which are not
within a loop are plainly wrong.  The evil thing is that bugs caused by this
will only surface with a low probability, and even less when the I/O and/or
network load is low, i.e. basically never when doing naive testing by hand.
:-P

As a bonus, we even fix the encoding issue by defining the whole protocol to
be UTF-8-based.