Better matching for hosts while joining the EC tables, Livestatus part.
This is basically handled like it is in CMC, but note one detail: The
initialization of some parts had to be moved to a later point in time in the
broker module lifecycle, because only then we know the hosts.

We should really introduce a kind of "resolver" object for stuff like this
and make it accessible to the parts which need it.  Currently this is all
global stuff :-P in Livestatus, while in CMC it is part of the World.
Passing the World (or even the whole Core to which it belongs) around just
to access a tiny bit of it needlessly increases coupling, so this is a bit
ugly.