2239 FIX Fixed exception in WATO snapins Folders, Tree of folders and Virtual Host Tree
An exception is thrown, if you want to use the WATO snapins Folders, Tree of folders and
Virtual Host Tree in a distributed monitoring setup. The master site is used as viewer
for the slave sites. If you create folders on slave sites and activate one of the snapins
at teh master site there is an error shown like No WATO folder ... . This have been fixed.