3720 The Event Console views are now supporting distributed setups
The Event Console was not really supporting distributed setups in the past. In case you have
a setup of multiple sites you always had to decide whether or not to setup a single, central,
EC or one EC per site and how to handle and configure monitoring and notifications.

The Event Console can now be integrated into distributed setups more easily. The standard
setup should now be a single Event Console per site which receives the events of all
hosts monitoringed with this site.

In case you have a central view site, the Event Console views show up the events of all Event
Console instances configured in the central site. You can manage all the events of the remote
Event Console just like managing hosts and services of the remote site.