3765 FIX oracle_asm_diskgroup: Made check more robust against invalid data received from agent
It may happen, at least with older version of the agent plugin, that the help text of SQLPlus
is sent in the agent sections. This could lead to ~50 services being created by accident. The
changed check now ignores these unexpected lines from the agent and only processes lines which
match the known format.